+++ addOns/reports/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Incorrect alert instances associated with alerts which have the same IDs (Issue 6873)
 
 ## [0.8.0] - 2021-10-07
 ### Added